requirements.txt
Django
djangorestframework
pandas
scikit-learn
odb
requests
numpy==1.23.5
loguru